feat: Add canvas element for alien image
This commit adds a canvas element for the alien image in the start/images.md file. The canvas element is created using the `GameWindowManager.addCanvasElement` function and the alien image is loaded and displayed on the canvas. This enhancement improves the functionality of the application by allowing users to initialize and add the image to the canvas.