feat: Load image textures in cache when a label is called
This commit updates the `StartLabel` class in the `start/images.md` file to load image textures in the cache when a label is called. The `Assets.load` function is used to load the `image1.png` and `image2.png` textures. This enhancement improves the performance of the application by preloading the textures, ensuring faster loading times when the label is displayed.