feat: Add TickerMove class for moving canvas elements
The TickerMove class has been added to enable moving canvas elements in both the x and y directions. This class provides options for setting the speed, destination, speed progression, and starting conditions of the movement. The default speed is 0.01, and the movement will start even if the canvas element does not have a texture. This enhancement improves the animation capabilities of the application and allows for more dynamic and interactive visual effects.