feat: Add Google Analytics tracking code to head configuration
This commit adds the Google Analytics tracking code to the head configuration in the `.vitepress/config.mts` file. The tracking code is included as two script tags, one for loading the Google Tag Manager script and another for initializing and configuring the tracking. This enhancement allows for the collection of website analytics data using Google Analytics.